+++ src/Tracing/Doctrine/DBAL/TracingDriverMiddleware.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace Sentry\SentryBundle\Tracing\Doctrine\DBAL;
 
-use Doctrine\DBAL\Driver as DriverInterface;
-use Doctrine\DBAL\Driver\Middleware as MiddlewareInterface;
+use Doctrine\DBAL\Driver;
+use Doctrine\DBAL\Driver\Middleware;
 use Sentry\State\HubInterface;
 
 /**
- * This middleware wraps a {@see DriverInterface} instance into one that
+ * This middleware wraps a {@see Driver} instance into one that
  * supports the distributed tracing feature of Sentry.
  */
-final class TracingDriverMiddleware implements MiddlewareInterface
+final class TracingDriverMiddleware implements Middleware
 {
     /**
      * @var HubInterface The current hub
@@ -32,7 +32,7 @@ final class TracingDriverMiddleware implements MiddlewareInterface
     /**
      * {@inheritdoc}
      */
-    public function wrap(DriverInterface $driver): DriverInterface
+    public function wrap(Driver $driver): Driver
     {
         return new TracingDriver($this->hub, $driver);
     }