+++ src/aliases.php
@@ -4,20 +4,14 @@ declare(strict_types=1);
 
 namespace Sentry\SentryBundle;
 
-use Doctrine\DBAL\Driver as DoctrineDriverInterface;
-use Doctrine\DBAL\Driver\DriverException as LegacyDriverExceptionInterface;
-use Doctrine\DBAL\Driver\Exception as DriverExceptionInterface;
 use Doctrine\DBAL\Driver\ExceptionConverterDriver as LegacyExceptionConverterDriverInterface;
 use Doctrine\DBAL\Driver\Middleware as DoctrineMiddlewareInterface;
-use Doctrine\DBAL\Driver\Result;
-use Doctrine\DBAL\Driver\Statement;
 use Sentry\SentryBundle\EventListener\ErrorListenerExceptionEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerControllerEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerRequestEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerResponseEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerTerminateEvent;
 use Sentry\SentryBundle\EventListener\SubRequestListenerRequestEvent;
-use Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility\DriverInterface;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility\ExceptionConverterDriverInterface;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility\MiddlewareInterface;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
@@ -94,24 +88,9 @@ if (version_compare(Kernel::VERSION, '4.3.0', '>=')) {
     }
 }
 
-if (interface_exists(Statement::class) && !interface_exists(Result::class)) {
-    /** @psalm-suppress UndefinedClass */
-    class_alias(Statement::class, Result::class);
-}
-
-if (interface_exists(DriverExceptionInterface::class) && !interface_exists(LegacyDriverExceptionInterface::class)) {
-    /** @psalm-suppress UndefinedClass */
-    class_alias(DriverExceptionInterface::class, LegacyDriverExceptionInterface::class);
-}
-
 if (!interface_exists(DoctrineMiddlewareInterface::class)) {
     /** @psalm-suppress UndefinedClass */
     class_alias(MiddlewareInterface::class, DoctrineMiddlewareInterface::class);
-}
-
-if (!interface_exists(DoctrineDriverInterface::class)) {
-    /** @psalm-suppress UndefinedClass */
-    class_alias(DriverInterface::class, DoctrineDriverInterface::class);
 }
 
 if (!interface_exists(LegacyExceptionConverterDriverInterface::class)) {