+++ tests/DependencyInjection/Fixtures/php/console_tracing_enabled.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+/** @var ContainerBuilder $container */
+$container->loadFromExtension('sentry', [
+    'tracing' => [
+        'console' => [
+            'excluded_commands' => [
+                'foo:bar',
+                'bar:foo',
+            ],
+        ],
+    ],
+]);